+++ packages/components/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.18.1](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.18.0...@jobber/components@4.18.1) (2023-05-19)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+# [4.18.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.17.1...@jobber/components@4.18.0) (2023-05-19)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
 ## [4.17.1](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.17.0...@jobber/components@4.17.1) (2023-05-18)
 
 **Note:** Version bump only for package @jobber/components